AS-IS) 폴더 구조 변경 TO-BE) 리팩토링
