+++ src/main/java/com/spring/sharepod/model/UserInfo.java
@@ -1,11 +1,9 @@
 package com.spring.sharepod.model;
 
-import com.spring.sharepod.v1.dto.response.*;
+import com.spring.sharepod.v1.dto.response.User.UserInfoResponseDto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
-import java.util.List;
 
 @AllArgsConstructor
 @NoArgsConstructor
