+++ src/main/java/com/spring/sharepod/model/UserOrder.java
@@ -1,6 +1,7 @@
 package com.spring.sharepod.model;
 
 import com.spring.sharepod.v1.dto.response.*;
+import com.spring.sharepod.v1.dto.response.User.UserReservation;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
