+++ src/main/java/com/spring/sharepod/v1/controller/NoticeRestController.java
@@ -5,9 +5,7 @@ import com.spring.sharepod.entity.User;
 import com.spring.sharepod.model.NoticeCount;
 import com.spring.sharepod.model.NoticeInfo;
 import com.spring.sharepod.model.Success;
-import com.spring.sharepod.v1.dto.response.Notice;
-import com.spring.sharepod.v1.dto.response.NoticeInfoList;
-import com.spring.sharepod.v1.dto.response.NoticeResponseDto;
+import com.spring.sharepod.v1.dto.response.notice.Notice;
 import com.spring.sharepod.v1.service.NoticeService;
 import com.spring.sharepod.v1.validator.TokenValidator;
 import lombok.RequiredArgsConstructor;
