+++ src/main/java/com/spring/sharepod/v1/controller/UserRestController.java
@@ -4,11 +4,13 @@ import com.spring.sharepod.entity.User;
 import com.spring.sharepod.model.*;
 import com.spring.sharepod.v1.dto.request.UserRequestDto;
 import com.spring.sharepod.v1.dto.response.*;
+import com.spring.sharepod.v1.dto.response.User.UserInfoResponseDto;
+import com.spring.sharepod.v1.dto.response.User.UserReservation;
+import com.spring.sharepod.v1.dto.response.User.UserResponseDto;
 import com.spring.sharepod.v1.service.AwsS3Service;
 import com.spring.sharepod.v1.service.UserService;
 import com.spring.sharepod.v1.validator.TokenValidator;
 import com.spring.sharepod.v1.validator.UserValidator;
-import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
