+++ src/main/java/com/spring/sharepod/v1/dto/response/Liked/LikedListResponseDto.java
@@ -1,4 +1,4 @@
-package com.spring.sharepod.v1.dto.response;
+package com.spring.sharepod.v1.dto.response.Liked;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
