+++ src/main/java/com/spring/sharepod/v1/dto/response/Reservation/ReservationGetDTO.java
@@ -1,4 +1,4 @@
-package com.spring.sharepod.v1.dto.response;
+package com.spring.sharepod.v1.dto.response.Reservation;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.*;
