+++ src/main/java/com/spring/sharepod/v1/dto/response/Reservation/ReservationNoticeList.java
@@ -1,4 +1,4 @@
-package com.spring.sharepod.v1.dto.response;
+package com.spring.sharepod.v1.dto.response.Reservation;
 
 
 import com.spring.sharepod.entity.User;
