+++ src/main/java/com/spring/sharepod/v1/dto/response/User/UserReservation.java
@@ -1,9 +1,10 @@
-package com.spring.sharepod.v1.dto.response;
+package com.spring.sharepod.v1.dto.response.User;
+
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.*;
 
-import java.time.LocalDateTime;
+import java.time.LocalDate;
 import java.util.Optional;
 
 @Getter
@@ -11,22 +12,22 @@ import java.util.Optional;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class MyBoardResponseDto {
+public class UserReservation {
     @JsonProperty(value="boardId")
     private Long id;
 
     @JsonProperty(value="boardTitle")
     private String title;
-
-    private String boardTag;
     private String boardRegion;
-
-    @JsonProperty(value="firstImg")
+    private String boardTag;
     private String firstImgUrl;
-
-    private LocalDateTime modifiedAt;
     private int dailyRentalFee;
+    private LocalDate startRental;
+    private LocalDate endRental;
     private String nickName;
+
     private String category;
+    private Long reservationId;
     private Optional<Boolean> isLiked;
+
 }
