+++ src/main/java/com/spring/sharepod/v1/dto/response/notice/Notice.java
@@ -1,4 +1,4 @@
-package com.spring.sharepod.v1.dto.response;
+package com.spring.sharepod.v1.dto.response.notice;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
