+++ src/main/java/com/spring/sharepod/v1/dto/response/notice/NoticeInfoList.java
@@ -1,4 +1,4 @@
-package com.spring.sharepod.v1.dto.response;
+package com.spring.sharepod.v1.dto.response.notice;
 
 
 import lombok.*;
