+++ src/main/java/com/spring/sharepod/v1/dto/response/notice/NoticeResponseDto.java
@@ -1,4 +1,4 @@
-package com.spring.sharepod.v1.dto.response;
+package com.spring.sharepod.v1.dto.response.notice;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
