+++ src/main/java/com/spring/sharepod/v1/repository/Board/BoardRepositoryCustom.java
@@ -2,6 +2,8 @@ package com.spring.sharepod.v1.repository.Board;
 
 import com.spring.sharepod.v1.dto.response.*;
 import com.spring.sharepod.v1.dto.response.Board.BoardAllResponseDto;
+import com.spring.sharepod.v1.dto.response.Board.MyBoardResponseDto;
+import com.spring.sharepod.v1.dto.response.User.UserReservation;
 import com.spring.sharepod.v1.repository.SearchForm;
 
 import java.util.List;
