+++ src/main/java/com/spring/sharepod/v1/repository/Reservation/ReservationRepositoryCustom.java
@@ -1,7 +1,7 @@
 package com.spring.sharepod.v1.repository.Reservation;
 
-import com.spring.sharepod.v1.dto.response.ReservationGetDTO;
-import com.spring.sharepod.v1.dto.response.ReservationNoticeList;
+import com.spring.sharepod.v1.dto.response.Reservation.ReservationGetDTO;
+import com.spring.sharepod.v1.dto.response.Reservation.ReservationNoticeList;
 
 import java.util.List;
